+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 - Added `Delivery` annotation
 - Added `DispatchManager` class. It will dispatch its deliverable object to setters (denoted by @Delivery) of a factory
 - Added `Deliverable` class, which contains a payload to be delivered
+- Added `workflow.v2.Pipeline`
+- Added `workflow.v2.stage`
+- Added `transformation.Factory[T]`
 
 ## 0.2.7 (2019-06-21)
 - Added `Conf` into `SparkRepositoryBuilder` and changed all the set methods